--- a/onnxruntime/contrib_ops/webgpu/quantization/matmul_nbits.h
+++ b/onnxruntime/contrib_ops/webgpu/quantization/matmul_nbits.h
@@ -31,6 +31,25 @@ class MatMulNBitsProgram final : public Program<MatMulNBitsProgram> {
   bool use_block32_;
 };
 
+
+class MatMulNBitsProgramPrefill final : public Program<MatMulNBitsProgramPrefill> {
+ public:
+  MatMulNBitsProgramPrefill(bool has_zero_points) : Program{"MatMulNBitsPrefill"},
+                                              has_zero_points_{has_zero_points} {
+  }
+
+  Status GenerateShaderCode(ShaderHelper& sh) const override;
+  WEBGPU_PROGRAM_DEFINE_UNIFORM_VARIABLES(
+    {"M", ProgramUniformVariableDataType::Uint32},
+    {"N", ProgramUniformVariableDataType::Uint32},
+    {"K", ProgramUniformVariableDataType::Uint32},
+    {"K4", ProgramUniformVariableDataType::Uint32},
+    {"K8", ProgramUniformVariableDataType::Uint32});
+
+ private:
+  bool has_zero_points_;
+};
+
 class MatMulNBits final : public WebGpuKernel {
  public:
   MatMulNBits(const OpKernelInfo& info) : WebGpuKernel(info) {