@guschmue 2024-12-10
just needs a fix for the macos build