@guschmue 2024-12-10
I see similar prefill improvements on Xe TGL, ~1.5x